pyproject.toml
[project]
name = "graphrag"
version = "0.1.0"
description = ""
authors = [
    {name = "Hedda G",email = "hedda.gressel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"

[tool.poetry]
name = "graphrag"
version = "0.1.0"
description = ""
authors = ["Hedda G <hedda.gressel@gmail.com>"]
readme = "README.md"
packages = [
    { include = "documentassistent" }
]

[tool.poetry.dependencies]
python = "^3.12"
langchain = ">=0.3.14,<0.4.0"
langchain-community = ">=0.3.19,<0.4.0"
langchain-ollama = ">=0.2.3,<0.3.0"
langchain-openai = ">=0.3.16,<0.4.0"
pytesseract = ">=0.3.13,<0.4.0"
openai = ">=1.76.2,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
loguru = ">=0.7.3,<0.8.0"
pydantic = "^2.5.3"
langfuse = "^2.60.7"
pypdf2 = "^3.0.1"
langgraph = "^0.5.4"
sqlalchemy = "^2.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"
mypy = "^1.8.0"
types-PyYAML = ">=6.0.12.20250516,<7.0.0"
types-requests = "*"
pytest = ">=8.3.5,<9.0.0"
pytest-mock = ">=3.14.0,<4.0.0"
pytest-cov = ">=6.1.1,<7.0.0"
types-pillow = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100",
    "D107",
    "ANN204",
    "D413",
    "RET504",
    "D401",
    "INP001",
    "PLE1205",
    "S101",
    "D104",
    "D203",
    "D212",
    "ANN401",
    "S603",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "PT001", "PT023"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=. --cov-report=xml"


